backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.3
Werkzeug==3.0.1
passlib==1.7.4
SQLAlchemy==2.0.23
